+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Table of Contents
 
+- [v0.77.0](#v0770)
+- [v0.76.1](#v0761)
 - [v0.76.0](#v0760)
 - [v0.75.1](#v0751)
 - [v0.75.0](#v0750)
@@ -95,6 +97,14 @@
 - [0.3.0](#030)
 - [0.2.0](#020)
 - [0.1.0](#010)
+
+## [v0.77.0]
+
+> Release date: 2026/07/15
+
+- Added Client APIs (Create, Delete, Get, List, Update) 
+  to interact with `AIModel`.
+  [#624](https://github.com/Kong/go-kong/pull/624)
 
 ## [v0.76.1]
 
@@ -1184,6 +1194,8 @@ authentication credentials in Kong.
   releases of Kong since every release of Kong is introducing breaking changes
   to the Admin API.
 
+[v0.77.0]: https://github.com/Kong/go-kong/compare/v0.76.1...v0.77.0
+[v0.76.1]: https://github.com/Kong/go-kong/compare/v0.76.0...v0.76.1
 [v0.76.0]: https://github.com/Kong/go-kong/compare/v0.75.1...v0.76.0
 [v0.75.1]: https://github.com/Kong/go-kong/compare/v0.75.0...v0.75.1
 [v0.75.0]: https://github.com/Kong/go-kong/compare/v0.74.0...v0.75.0